fifa_tournament/requirements.txt
amqp==2.2.1
aspy.yaml==0.3.0
astroid==1.5.3
autopep8==1.3.2
bandit==1.4.0
billiard==3.5.0.3
boltons==16.1.1
brotlipy==0.7.0
cached-property==1.3.0
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
coverage==4.4.1
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.4
django-log-request-id==1.3.2
django-model-utils==3.0.0
django-naomi==0.8
django-webpack-loader==0.5.0
django.js-vinta==0.8.2.dev2
djangorestframework==3.6.4
dodgy==0.1.9
dparse==0.2.1
fixmydjango==0.2
gitdb2==2.0.2
GitPython==2.1.5
gunicorn==19.7.1
identify==1.0.5
idna==2.6
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.3.2
nodeenv==1.2.0
opbeat==3.5.2
packaging==16.8
pbr==3.1.1
pep8-naming==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
pre-commit==0.17.0
prompt-toolkit==1.0.15
prospector==0.12.7
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycodestyle==2.0.0
pycparser==2.18
pydocstyle==2.0.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.2
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
python-decouple==3.1
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requirements-detector==0.5.2
safety==1.5.1
setoptconf==0.2.0
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.3
snowballstemmer==1.2.1
stevedore==1.25.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
vulture==0.26
wcwidth==0.1.7
whitenoise==3.3.0
wrapt==1.10.11
